States are saved as automanaged objects
Previously States were saved by the Simulation object as raw pointers. There was no need for this and so they are now stored as value objects and so have their lifetime determined by their scope. When accessing states an observing pointer (raw pointer) is returned. Ideally it'd be a reference but I can't get that to work